[track] split TracksBuilder and use pimpl to cut dependency to lemon
