+++ src/aliceVision/track/CMakeLists.txt
@@ -2,11 +2,13 @@
 set(tracks_files_headers
   Track.hpp
   TracksBuilder.hpp
+  tracksUtils.hpp
 )
 
 # Sources
 set(tracks_files_sources
   TracksBuilder.cpp
+  tracksUtils.cpp
 )
 
 alicevision_add_library(aliceVision_track
