+++ src/aliceVision/track/Track.hpp
@@ -26,6 +26,38 @@ namespace track {
 
 using namespace aliceVision::matching;
 
+using FeatureId = std::pair<feature::EImageDescriberType, std::size_t>;
+
+
+/**
+ * @brief KeypointId is a unique ID for a feature in a view.
+ */
+struct KeypointId
+{
+  KeypointId(){}
+  KeypointId(feature::EImageDescriberType type, std::size_t index)
+    : descType(type)
+    , featIndex(index)
+  {}
+
+  bool operator<(const KeypointId& other) const
+  {
+    if(descType == other.descType)
+      return featIndex < other.featIndex;
+    return descType < other.descType;
+  }
+
+  feature::EImageDescriberType descType = feature::EImageDescriberType::UNINITIALIZED;
+  std::size_t featIndex = 0;
+};
+
+inline std::ostream& operator<<(std::ostream& os, const KeypointId& k)
+{
+    os << feature::EImageDescriberType_enumToString(k.descType) << ", " << k.featIndex;
+    return os;
+}
+
+
 /**
  * @brief A Track is a feature visible accross multiple views.
  * Tracks are generated by the fusion of all matches accross all images.
@@ -33,7 +65,7 @@ using namespace aliceVision::matching;
 struct Track
 {
   /// Data structure to store a track: collection of {ViewId, FeatureId}
-  typedef stl::flat_map<std::size_t, std::size_t> FeatureIdPerView;
+  using FeatureIdPerView = stl::flat_map<std::size_t, std::size_t>;
 
   Track() {}
 
@@ -44,8 +76,8 @@ struct Track
 };
 
 /// A track is a collection of {trackId, Track}
-typedef stl::flat_map<std::size_t, Track> TracksMap;
-typedef std::vector<std::size_t> TrackIdSet;
+using TracksMap = stl::flat_map<std::size_t, Track>;
+using TrackIdSet = std::vector<std::size_t>;
 
 /**
  * @brief Data structure that contains for each features of each view, its corresponding cell positions for each level of the pyramid, i.e.
@@ -61,13 +93,13 @@ typedef std::vector<std::size_t> TrackIdSet;
  * and we go on for increasing values of l so that e.g. the first cell of the pyramid at l=2 has position K^2, the second K^2 + 1 etc...
  * So in general the i-th cell of the pyramid at level l has position P= \sum_{j=1...l-1} K_j^2 + i
  */
-typedef stl::flat_map<std::size_t, stl::flat_map<std::size_t, std::size_t> > TracksPyramidPerView;
+using TracksPyramidPerView = stl::flat_map<std::size_t, stl::flat_map<std::size_t, std::size_t> >;
 
 /**
  * @brief TracksPerView is a list of visible track ids for each view.
  * TracksPerView contains <viewId, vector<trackId>>
  */
-typedef stl::flat_map<std::size_t, TrackIdSet > TracksPerView;
+using TracksPerView = stl::flat_map<std::size_t, TrackIdSet >;
 
 
 } // namespace track
