+++ src/aliceVision/track/track_test.cpp
@@ -6,6 +6,7 @@
 // You can obtain one at https://mozilla.org/MPL/2.0/.
 
 #include "aliceVision/track/TracksBuilder.hpp"
+#include "aliceVision/track/tracksUtils.hpp"
 #include "aliceVision/matching/IndMatch.hpp"
 
 #include <vector>
@@ -194,11 +195,11 @@ BOOST_AUTO_TEST_CASE(Track_GetCommonTracksInImages)
     map_tracksPerView[20].push_back(6);
 
     std::set<std::size_t> set_visibleTracks;
-    tracksUtilsMap::getCommonTracksInImages(set_imageIndex, map_tracksPerView, set_visibleTracks);
+    getCommonTracksInImages(set_imageIndex, map_tracksPerView, set_visibleTracks);
     BOOST_CHECK_EQUAL(base.size(), set_visibleTracks.size());
     set_visibleTracks.clear();
     // test non-existing view index
-    tracksUtilsMap::getCommonTracksInImages({15, 50}, map_tracksPerView, set_visibleTracks);
+    getCommonTracksInImages({15, 50}, map_tracksPerView, set_visibleTracks);
     BOOST_CHECK(set_visibleTracks.empty());
   }
   {
@@ -223,7 +224,7 @@ BOOST_AUTO_TEST_CASE(Track_GetCommonTracksInImages)
     map_tracksPerView[20].push_back(6);
 
     std::set<std::size_t> set_visibleTracks;
-    tracksUtilsMap::getCommonTracksInImages(set_imageIndex, map_tracksPerView, set_visibleTracks);
+    getCommonTracksInImages(set_imageIndex, map_tracksPerView, set_visibleTracks);
     BOOST_CHECK_EQUAL(base.size(), set_visibleTracks.size());
   }
 }
